fix(warden): move `api` pkg to match `ignite` expectations
