+++ warden/x/warden/module/autocli.go
@@ -3,7 +3,7 @@ package warden
 import (
 	autocliv1 "cosmossdk.io/api/cosmos/autocli/v1"
 
-	modulev1 "github.com/warden-protocol/wardenprotocol/warden/api/warden/warden"
+	modulev1 "github.com/warden-protocol/wardenprotocol/api/warden/warden"
 )
 
 // AutoCLIOptions implements the autocli.HasAutoCLIConfig interface.
